image_generator/filter_top_k_images_and_export/requirements.txt
pandas>=1.3.0
openpyxl>=3.1.0
Pillow>=9.0.0
